castle treasury int -> long
git-svn-id: http://svn.l2jserver.com/trunk@3152 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/actor/instance/L2MerchantInstance.java
@@ -246,7 +246,7 @@ public class L2MerchantInstance extends L2NpcInstance
         price *= cost[val - 1];
         int time = ridetime[val - 1];
 
-        if (!player.reduceAdena("Rent", (int) price, player.getLastFolkNPC(), true)) return;
+        if (!player.reduceAdena("Rent", (long) price, player.getLastFolkNPC(), true)) return;
 
         player.mount(petId, 0, false);
         SetupGauge sg = new SetupGauge(3, time*1000);
